requirements.txt
numpy>=1.25,<3.0
pandas>=2.0,<3.0
faiss-cpu>=1.13.2
sentence-transformers>=2.2
google-generativeai>=0.8
python-dotenv>=1.0
tqdm
python-dotenv>=1.0
fastapi>=0.110
uvicorn>=0.30

